+++ src/soprasteria/code24h/marabunta/strategie/fourmiliere/StrategieFourmiliereBasique.java
@@ -13,7 +13,7 @@ public class StrategieFourmiliereBasique implements StrategieFourmiliere {
 	@Override
 	public void cogite(Fourmiliere fourmiliere) {
 		int fourmisInterieur = fourmiliere.getNbFourmis();
-		short fourmisExterieur = fourmiliere.getMemories()[0];
+		Integer fourmisExterieur = fourmiliere.getMemories()[0];
 		System.out.println(": > interieur : " + fourmisInterieur + " exterieur : " + fourmisExterieur);
 		if(fourmisExterieur < 20 && fourmisInterieur > 0) {
 			actionsFourmiliere.SetMemory(++fourmisExterieur);
